Combine the 1D and 2D plot elements into a single element
Review: @joshmutus
