+++ client-js/app/elements/elements.html
@@ -37,8 +37,7 @@
 <link rel="import" href="labrad-breadcrumbs.html">
 <link rel="import" href="labrad-grapher.html">
 <link rel="import" href="labrad-grapher-live.html">
-<link rel="import" href="labrad-plot1d.html">
-<link rel="import" href="labrad-plot2d.html">
+<link rel="import" href="labrad-plot.html">
 <link rel="import" href="labrad-manager.html">
 <link rel="import" href="labrad-nodes.html">
 <link rel="import" href="labrad-registry.html">
