+++ client-js/app/elements/labrad-grapher-live.html
@@ -22,12 +22,11 @@
     <style>
       :host {
         width: 100%;
-        flex-grow: 0.33;
         display: flex;
         flex-direction: column;
       }
       #plot {
-        flex-grow: 1;
+        height: 400px;
         display: flex;
       }
     </style>
