--- client-js/app/elements/labrad-plot2d.html
@@ -1,11 +1,0 @@
-<dom-module id="labrad-plot2d">
-  <style>
-    :host {
-      height: 100%;
-      display: flex;
-    }
-  </style>
-  <template>
-    <div id="area" class="flex"></div>
-  </template>
-</dom-module>
